Refactor: 기타 함수 분리.
