+++ src/basic/components/CartItems.js
@@ -1,5 +1,4 @@
 const CartItems = () => {
-  // Cart items container
   const $cartItems = document.createElement('div');
   $cartItems.id = 'cart-items';
